requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gymnasium==0.29.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
mss==9.0.2
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
stable_baselines3==2.3.2
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
